fixed lint too many functions error for YouTubeClient
